Add files via upload
I have worked on Predicting House Prices using Machine Learning. In this task, we performed a series of steps to develop a linear regression model for predicting house prices based on a dataset. First, we loaded and explored the dataset to ensure there were no missing values. We selected three key features: the size of the house in square feet, the number of bedrooms, and the number of bathrooms. These features were used to train the model. We then split the data into training and testing sets to evaluate the model's performance. A linear regression model was trained using the training set, and predictions were made on the test set. Finally, we evaluated the model using the Root Mean Squared Error (RMSE) metric, which provided an indication of the average prediction error in dollars.